tests: implement integration test
An integration test to execute the project from end to end has been
added. It requires a local DNS resolver with the expected registries
running on port 5353. CoreDNS is the recommended server, the QA workflow
shows how to run it.

Some of the source code was modified to make testing easier, mainly
changing function signatures to take string slices instead of os.Args
and io.Writer instead of hardcoding os.Stdout.

Closes #2.